Fix for pathing Issues
Issue that could cause files not to load by using Linux.

Also, why I didn't do this sooner is because I was working in a windows environment and was not even thinking of Linux users, always learning.

This bot project was written at a time when I was still and am learning, if I had my time again, I would do it a different way!

-----

CHANGED Lang file to be more specific to this bot.
DISABLED ready Status changing [You can put it back, still works fine. Just uncomment the code and it will be good to go.]